Text refactored to TextObject to avoid colliding with another library

+++ src/main/java/us/mudkip989/mods/cge/utils/Meths.java
@@ -46,7 +46,7 @@ public class Meths {
 
     }
 
-    public static Location shiftLocationForwards(Location loc, Integer amount){
+    public static Location shiftLocationForwards(Location loc, Float amount){
         Vector dir = loc.getDirection();
         dir.multiply(amount);
         loc.add(dir);
